Initial Commit. Problem with cert renewal conf.
The basic functionality already works, but the required
config for certbot to renew the certificates was generated
empty. The first issue that needs to be fixed.

After that, make sure that the scripts are parameterized
nicer, try to auto-detect some settings, like if amd64
is present. Also add some docs to the README!